refactor: simplify the Generate component by modularizing the form and header into separate components; enhance file processing logic and improve state management for better maintainability
