+++ app/(dashboard)/generate/components/generator-form.tsx
@@ -1,0 +1,211 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { Card, CardContent } from "@/components/ui/card";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { Sparkles } from "lucide-react";
+import { useEffect } from "react";
+import { useForm } from "react-hook-form";
+import { GeneratorFormSchema, type GeneratorFormData } from "../schema";
+import type { FileProcessingState, GeneratorForm } from "../types";
+import { FileUpload } from "./file-upload";
+
+type GeneratorFormProps = {
+  onSubmit: (data: GeneratorForm) => Promise<void>;
+  selectedFiles: File[];
+  setSelectedFiles: (files: File[] | ((prev: File[]) => File[])) => void;
+  onRecognizedText: (text: string, fileId: string) => void;
+  onTextRemoved: (fileId: string) => void;
+  fileProcessingStates: FileProcessingState;
+  setFileProcessingStates: (
+    states:
+      | FileProcessingState
+      | ((prev: FileProcessingState) => FileProcessingState)
+  ) => void;
+  processedFiles: Set<string>;
+  setProcessedFiles: (
+    files: Set<string> | ((prev: Set<string>) => Set<string>)
+  ) => void;
+  isRecognizing: boolean;
+};
+
+export function GeneratorForm({
+  onSubmit,
+  selectedFiles,
+  setSelectedFiles,
+  onRecognizedText,
+  onTextRemoved,
+  fileProcessingStates,
+  setFileProcessingStates,
+  processedFiles,
+  setProcessedFiles,
+  isRecognizing,
+}: GeneratorFormProps) {
+  const form = useForm<GeneratorFormData>({
+    resolver: zodResolver(GeneratorFormSchema),
+    defaultValues: {
+      title: "",
+      subject: "",
+      level: "",
+      files: [],
+    },
+  });
+
+  const isInputFilled =
+    form.watch("title") &&
+    form.watch("subject") &&
+    form.watch("level") &&
+    selectedFiles.length > 0;
+
+  const handleFormSubmit = async (data: GeneratorFormData) => {
+    try {
+      const formFields: GeneratorForm = {
+        option: "files",
+        title: data.title,
+        subject: data.subject,
+        level: data.level,
+        files: data.files,
+      };
+      await onSubmit(formFields);
+    } catch (error: unknown) {
+      console.error("Error submitting form: ", error);
+    }
+  };
+
+  const handleFilesChange = (files: File[]) => {
+    form.setValue("files", files);
+  };
+
+  // Watch form values for validation
+  useEffect(() => {
+    const subscription = form.watch(() => {
+      // Trigger validation when form values change
+    });
+    return () => subscription.unsubscribe();
+  }, [form]);
+
+  return (
+    <Card className="border-0 shadow-lg bg-white/70 backdrop-blur-sm max-w-4xl mx-auto w-full">
+      <CardContent className="p-8">
+        <Form {...form}>
+          <form
+            onSubmit={form.handleSubmit(handleFormSubmit)}
+            className="flex flex-col gap-6"
+          >
+            {/* Title Field */}
+            <FormField
+              control={form.control}
+              name="title"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel className="text-gray-800 font-semibold">
+                    Titre
+                  </FormLabel>
+                  <FormControl>
+                    <Input
+                      placeholder="Les équations du second degré"
+                      className="h-12 border-blue-200/60 focus:border-blue-600 focus:ring-blue-600/20 bg-white/80 backdrop-blur-sm"
+                      {...field}
+                    />
+                  </FormControl>
+                  <FormDescription className="text-gray-600">
+                    Renseigne le titre de ton cours.
+                  </FormDescription>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+
+            <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
+              {/* Subject Field */}
+              <FormField
+                control={form.control}
+                name="subject"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel className="text-gray-800 font-semibold">
+                      Matière
+                    </FormLabel>
+                    <FormControl>
+                      <Input
+                        placeholder="Mathématiques"
+                        className="h-12 border-blue-200/60 focus:border-blue-600 focus:ring-blue-600/20 bg-white/80 backdrop-blur-sm"
+                        {...field}
+                      />
+                    </FormControl>
+                    <FormDescription className="text-gray-600">
+                      La matière de ton cours.
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+
+              {/* Level Field */}
+              <FormField
+                control={form.control}
+                name="level"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel className="text-gray-800 font-semibold">
+                      Niveau
+                    </FormLabel>
+                    <FormControl>
+                      <Input
+                        placeholder="Seconde"
+                        className="h-12 border-blue-200/60 focus:border-blue-600 focus:ring-blue-600/20 bg-white/80 backdrop-blur-sm"
+                        {...field}
+                      />
+                    </FormControl>
+                    <FormDescription className="text-gray-600">
+                      Le niveau d'étude de ton cours.
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+            </div>
+
+            {/* File Upload Field */}
+            <FormField
+              control={form.control}
+              name="files"
+              render={() => (
+                <FileUpload
+                  selectedFiles={selectedFiles}
+                  setSelectedFiles={setSelectedFiles}
+                  onFilesChange={handleFilesChange}
+                  onRecognizedText={onRecognizedText}
+                  onTextRemoved={onTextRemoved}
+                  fileProcessingStates={fileProcessingStates}
+                  setFileProcessingStates={setFileProcessingStates}
+                  processedFiles={processedFiles}
+                  setProcessedFiles={setProcessedFiles}
+                />
+              )}
+            />
+
+            <Button
+              disabled={!isInputFilled || isRecognizing}
+              type="submit"
+              className="h-12 bg-gradient-to-r from-blue-600 to-indigo-600 hover:from-blue-700 hover:to-indigo-700 text-white font-semibold rounded-xl shadow-lg hover:shadow-xl transition-all duration-200 disabled:opacity-50"
+            >
+              <Sparkles className="w-5 h-5 mr-2" />
+              Lancer la génération
+            </Button>
+          </form>
+        </Form>
+      </CardContent>
+    </Card>
+  );
+}
